Unify memory allocators for SPIRam
