+++ src/apps/lilcatalog/lilcatalog.cpp
@@ -3,7 +3,7 @@
 #include <WiFi.h>
 #include <lilka/config.h>
 
-#include "keira/utils/json.h"
+#include "keira/utils/mem.h"
 
 LilCatalogApp::LilCatalogApp() : App(K_S_LILCATALOG_APP), currentEntry{}, iconBuffer{}, downloadBuffer{} {
     setktStackSize(16384);
