+++ src/apps/weather/weather.cpp
@@ -10,7 +10,7 @@
 // Services:
 #include "services/clock/clock.h"
 // Utils:
-#include "keira/utils/json.h"
+#include "keira/utils/mem.h"
 #include "keira/utils/string.h"
 
 // Icons:
